build: moved all stylesheets into separate folder
