+++ client/components/Services.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './Services.css';
+import '../stylesheets/Services.css';
 
 const Services = () => {
   return (
