+++ client/stylesheets/App.css
@@ -46,6 +46,8 @@ h6 {
 #__next {
   isolation: isolate;
 }
+
 .wrapper {
   max-width: 1520px;
+  margin: 0 auto;
 }
