+++ client/stylesheets/Founder.css
@@ -1,10 +1,10 @@
 #founder {
+  width: 100%;
   background: #000;
   color: #fff;
   display: flex;
   justify-content: center;
   align-items: center;
-  margin-left: -172px;
 }
 .quote {
   max-width: 26%;
